Bug Fix: Restricted areas for non-users and costumers added.

+++ scripts/deleteProduct-submit.php
@@ -12,7 +12,20 @@ $conn = new mysqli($servername, $username, $password, $dbname);
 
 $delete= $_POST['delete'];
 
-delete_product($conn, $delete);
+session_start();
+    if (isset($_SESSION['username']))
+    {
+        $loggedin = TRUE;
+        $role = $_SESSION['role'];
+    }
+    else $loggedin = FALSE;
+
+    if ($loggedin && $role='admin'){
+        delete_product($conn, $delete);
+    } else {
+        echo "Restricted area. Please <a href='../index.php'>click here</a> to go back to the Homepage.";
+    }
+
 
 
 /** @fn     'Delete Product'
